Server/requirements.txt
openai==1.11.1
termcolor==2.3.0
requests==2.31.0
opencv-python==4.9.0.80
pyshine==0.0.9
google-search-results==2.4.2
pinecone-client==2.2.4
Pillow==10.0.0
torchvision==0.17.0
torch==2.2.0
transformers~=4.39.3
serpapi~=0.1.5
numpy~=1.26.4
pandas~=2.2.1
python-dotenv~=1.0.1